change toggle button file wrap the app by suspense and delete it from the index file
