+++ client/package.json
@@ -11,7 +11,6 @@
     "i18next-http-backend": "^1.0.15",
     "i18next-xhr-backend": "^3.2.2",
     "prop-types": "^15.7.2",
-    "rc-switch": "^3.2.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-i18next": "^11.7.0",
