+++ client/public/locales/ar.json
@@ -1,5 +1,5 @@
 {
-  "main.header.welcome": "مرحبا",
+  "logo": "مرحبا",
   "main.page.title": "{{title}} اللغة العربية",
   "main.page.section.languages.title": "اللغة",
   "main.page.section.languages.subtitle": "ابدا من هنا"
