+++ client/src/components/Layout/Header.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import LanguageToggle from '../ToggleButton';
+import LanguageToggle from '../LanguageToggle';
 
 const Header = () => {
   const { t } = useTranslation();
 
   return (
-    <div className="bg-green-600 flex justify-between p-6">
-      <h1>{t('logo')}</h1>
+    <div className="bg-primary flex justify-between p-6">
+      <span className="text-white font-bold">{t('logo')}</span>
       <LanguageToggle />
     </div>
   );
