+++ client/src/components/index.js
@@ -2,6 +2,6 @@ export { default as Layout, Header, NavBar } from './Layout';
 export { BarChart, LineChart, PieChart } from './Charts';
 export { default as BillsTable } from './BillsTable';
 export { default as BillTypes } from './BillTypes';
-export { default as ToggleButton } from './ToggleButton';
+export { default as LanguageToggle } from './LanguageToggle';
 export { default as ToggleContainer } from './ToggleContainer';
 export { default as Tips } from './Tips';
